Improved props in button.
Properties should be obtained from props at top of the component and instead of using props.[Property Name], we should use these obtained properties to make the code more readable and manageable.
Fixed the React: "No validations for props" by adding validations for expected properties.

setLoading function from props was not used in the code, so it is removed.